refactor: 외부 영역 클릭 시 닫히는 함수 커스텀 훅으로 분리
