+++ hooks/use-click-outside.ts
@@ -1,0 +1,23 @@
+import { useEffect, useRef } from "react";
+
+function useClickOutside<T extends HTMLElement>(
+  handler: (event: MouseEvent) => void,
+) {
+  const ref = useRef<T>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        handler(event);
+      }
+    };
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [handler]);
+
+  return ref;
+}
+
+export default useClickOutside;
